save cart data by dataStore instead of calling api

+++ app/src/main/java/com/example/foodapp/data/model/CheckoutDetails.kt
@@ -1,8 +1,8 @@
 package com.example.foodapp.data.model
 
 data class CheckoutDetails(
-    val deliveryFee: Double,
-    val subTotal: Double,
-    val tax: Double,
-    val totalAmount: Double
+    val deliveryFee: Float,
+    val subTotal: Float,
+    val tax: Float,
+    val totalAmount: Float
 )
